完成对LogsPage的改造
